feat: access adapter removal sha2 (#2375)
Resolves INT-5950.

+++ extensions/sha2/circuit/src/sha2_chips/tests.rs
@@ -136,9 +136,9 @@ fn set_and_execute_single_block<RA: Arena, C: Sha2Config, E: PreflightExecutor<F
     let dst_ptr = gen_pointer(rng, 4);
     let state_ptr = gen_pointer(rng, 4);
     let input_ptr = gen_pointer(rng, 4);
-    tester.write(1, rd, dst_ptr.to_le_bytes().map(F::from_canonical_u8));
-    tester.write(1, rs1, state_ptr.to_le_bytes().map(F::from_canonical_u8));
-    tester.write(1, rs2, input_ptr.to_le_bytes().map(F::from_canonical_u8));
+    tester.write(1, rd, (dst_ptr as u32).to_le_bytes().map(F::from_canonical_u8));
+    tester.write(1, rs1, (state_ptr as u32).to_le_bytes().map(F::from_canonical_u8));
+    tester.write(1, rs2, (input_ptr as u32).to_le_bytes().map(F::from_canonical_u8));
 
     let default_message = get_random_message(rng, C::BLOCK_U8S);
     let message = message.unwrap_or(&default_message);
@@ -245,9 +245,9 @@ fn set_and_execute_full_message<RA: Arena, C: Sha2Config + 'static, E: Preflight
     let state_ptr = gen_pointer(rng, 4);
     let dst_ptr = state_ptr;
     let input_ptr = gen_pointer(rng, 4);
-    tester.write(1, rd, dst_ptr.to_le_bytes().map(F::from_canonical_u8));
-    tester.write(1, rs1, state_ptr.to_le_bytes().map(F::from_canonical_u8));
-    tester.write(1, rs2, input_ptr.to_le_bytes().map(F::from_canonical_u8));
+    tester.write(1, rd, (dst_ptr as u32).to_le_bytes().map(F::from_canonical_u8));
+    tester.write(1, rs1, (state_ptr as u32).to_le_bytes().map(F::from_canonical_u8));
+    tester.write(1, rs2, (input_ptr as u32).to_le_bytes().map(F::from_canonical_u8));
 
     // initial state as little-endian words
     let initial_state: Vec<u8> = C::get_h()
